Move client/new_service.rs -> client/connect.rs
