+++ src/client/mod.rs
@@ -1,7 +1,7 @@
 mod background;
-mod new_service;
+mod connect;
 mod connection;
 
 pub use self::background::Background;
-pub use self::new_service::{Connect, ConnectFuture, ConnectError};
+pub use self::connect::{Connect, ConnectFuture, ConnectError};
 pub use self::connection::{Connection, ResponseFuture, Error};
